requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# 資料庫
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
pgvector==0.2.3

# Redis
redis==5.0.1

# HTTP 客戶端
httpx==0.25.1
requests==2.31.0

# LangChain & AI
langchain==0.0.335
langchain-google-genai==0.0.5
langgraph==0.0.20
google-generativeai==0.3.1

# 認證與安全 (新增)
bcrypt==4.1.1
pyjwt==2.8.0
python-jose[cryptography]==3.3.0

# 工具
python-dotenv==1.0.0
pydantic-settings==2.1.0